requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
bitcoin-python==0.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
confusable-homoglyphs==3.3.1
coverage==7.6.10
cryptography==44.0.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.5
django-paypal==2.1
django-recaptcha==4.0.0
django-registration==5.1.0
django-taggit==6.1.0
djangorestframework==3.15.2
enum34==1.1.10
future==1.0.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
html2markdown==0.1.7
html2text==2024.2.26
httplib2==0.22.0
idna==3.10
isodate==0.7.2
lxml==5.3.0
misaka==2.1.1
mock==5.1.0
oauthlib==3.2.2
outcome==1.3.0.post0
pillow==11.1.0
psycopg==3.2.4
pycparser==2.22
pycryptodomex==3.21.0
Pygments==2.19.1
pyjwkest==1.4.2
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.28.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.6
splinter==0.21.0
sqlparse==0.5.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
Unipath==1.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
xxhash==3.5.0
